+++ README.md
@@ -3,6 +3,8 @@
 `VastVoid` is a small Pygame project that displays a world of star systems.
 You can control a ship and travel across a plane populated by stars with
 orbiting planets.
+The core game classes now live in separate modules within `src` and are
+initialized from a single entry point (`main.py`).
 
 ## Running the game
