+++ src/config.py
@@ -1,0 +1,13 @@
+WINDOW_WIDTH, WINDOW_HEIGHT = 800, 600
+BACKGROUND_COLOR = (0, 0, 20)  # near black
+
+SHIP_COLOR = (255, 255, 255)
+SHIP_SIZE = 20
+SHIP_ACCELERATION = 300  # pixels per second squared
+SHIP_FRICTION = 0.92  # velocity retained each frame
+
+ORBIT_SPEED_FACTOR = 0.005  # base factor used to calculate orbital speed
+
+SECTOR_WIDTH = 2000
+SECTOR_HEIGHT = 2000
+GRID_SIZE = 3
